propConfig wasn't returned for stuff nested in compound props. Fixed.
The reason it went unnoticed is that for now this caused only one thing to fail when something was nested in a coumpound prop: the value wasn't sanitized. The other thing that could've failed, but didn't, because we don't have that functionality yet, is setting the value of compound props as a whole, since the propConfig?.type === 'compound' check would've always been false because of the undefined propConfig.